@@ -1,22 +1,21 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.IO;
+﻿using System.Globalization;
 
 namespace paracobNET
 {
     public static class LabelIO
     {
-        public static OrderedDictionary<ulong, string> GetHashStringDict(string filepath)
+        public static OrderedDictionary<Hash40, string> GetHashStringDict(string filepath)
         {
-            OrderedDictionary<ulong, string> labels = new OrderedDictionary<ulong, string>();
+            OrderedDictionary<Hash40, string> labels = new OrderedDictionary<Hash40, string>();
             foreach (var line in File.ReadAllLines(filepath))
             {
                 string[] splits = line.Split(',');
                 try
                 {
                     if (splits[0].Substring(0, 2) == "0x")
-                        labels.Add(ulong.Parse(splits[0].Substring(2), NumberStyles.HexNumber), splits[1]);
+                        labels.Add(
+                            new Hash40(ulong.Parse(splits[0].Substring(2),
+                            NumberStyles.HexNumber)), splits[1]);
                     else throw new InvalidDataException();
                 }
                 catch (Exception e)
@@ -30,16 +29,18 @@ namespace paracobNET
             return labels;
         }
 
-        public static OrderedDictionary<string, ulong> GetStringHashDict(string filepath)
+        public static OrderedDictionary<string, Hash40> GetStringHashDict(string filepath)
         {
-            OrderedDictionary<string, ulong> labels = new OrderedDictionary<string, ulong>();
+            OrderedDictionary<string, Hash40> labels = new OrderedDictionary<string, Hash40>();
             foreach (var line in File.ReadAllLines(filepath))
             {
                 string[] splits = line.Split(',');
                 try
                 {
                     if (splits[0].Substring(0, 2) == "0x")
-                        labels.Add(splits[1], ulong.Parse(splits[0].Substring(2), NumberStyles.HexNumber));
+                        labels.Add(splits[1], new Hash40(
+                            ulong.Parse(splits[0].Substring(2),
+                            NumberStyles.HexNumber)));
                     else throw new InvalidDataException();
                 }
                 catch (Exception e)
@@ -53,11 +54,11 @@ namespace paracobNET
             return labels;
         }
 
-        public static void WriteLabels(string filepath, OrderedDictionary<ulong, string> labels)
+        public static void WriteLabels(string filepath, OrderedDictionary<Hash40, string> labels)
         {
             List<string> lines = new List<string>();
             foreach (var label in labels)
-                lines.Add($"0x{label.Key.ToString("x10")},{label.Value}");
+                lines.Add($"0x{label.Key},{label.Value}");
             File.WriteAllLines(filepath, lines);
         }
     }
